trainer/requirements.txt
numpy
matplotlib
pandas
seaborn
keras
scikit-learn
python-dotenv
jupyter
spotipy
librosa
music21
midi2audio
madmom
tqdm
tensorflow_datasets
pyaudio
tensorflowjs
pretty_midi